+++ base/exports.jl
@@ -964,7 +964,6 @@ export
     expand,
     gensym,
     macroexpand,
-    macroexpand1,
     @macroexpand1,
     @macroexpand,
     parse,