+++ NEWS.md
@@ -100,7 +100,7 @@ Library improvements
     `ntuple`, `Base.literal_pow`, `sqrtm`, `lufact`, `lufact!`, `qrfact`, `qrfact!`,
     `cholfact`, `cholfact!`, `_broadcast!`, `reshape`, `cat` and `cat_t`.
 
-  * A new `macroexpand1` function and `@macroexpand1` macro for non recursive macro expansion ([#21662]).
+  * A new `@macroexpand1` macro for non recursive macro expansion ([#21662]).
 
 Compiler/Runtime improvements
 -----------------------------